requirements.txt
django==4.2.11
psycopg2-binary==2.9.9
ipython==8.15.0
pillow==10.2.0
djangorestframework==3.15.0
django-filter==24.1
python-dotenv==1.0.1
djangorestframework-simplejwt==5.3.1
coverage==7.4.4
pytest==8.1.1
django-cors-headers==4.3.1
requests==2.31.0
stripe==8.8.0
django-celery-beat==2.6.0
celery==5.3.6
eventlet==0.35.2
redis==5.1.0b4
drf-yasg==1.21.7
django-cors-headers==4.3.1
flake8==7.0.0
